feat: extract RouteCache and apply keep-alive caching to TreeRouter
Move the keep-alive page cache out of Router into a reusable
<RouteCache> component and wire it into TreeRouter too (default cache of
10, 0 to disable), so documentation pages keep their state across
back/forward navigation. TreeProvider stays above the cache so the
flattened tree map is shared and stable; each cached page carries its
own frozen <Meta> context.
