+++ modules/ui/router/index.ts
@@ -1,5 +1,6 @@
 export * from "./Navigation.js";
 export * from "./NavigationContext.js";
 export * from "./NavigationStore.js";
+export * from "./RouteCache.js";
 export * from "./Router.js";
 export * from "./Routes.js";
